requirements.txt
apache-airflow==2.8.1
pandas>=1.3.0
numpy>=1.21.0
pendulum
apache-airflow-providers-google
apache-airflow-providers-amazon
google-cloud-bigquery
